Routes files renaming for less headaches when working in IDE
