+++ server/src/bootstrap/router/routes/index.js
@@ -1,7 +1,7 @@
-const { ideasRoutes } = require('./api/ideas')
-const { boardsRoutes } = require('./api/boards')
-const { usersRoutes } = require('./api/users')
-const { viewsRoutes } = require('./views/views')
+const { ideasRoutes } = require('./api/ideasRoutes')
+const { boardsRoutes } = require('./api/boardsRoutes')
+const { usersRoutes } = require('./api/usersRoutes')
+const { viewsRoutes } = require('./views/viewsRoutes')
 
 exports.default = {
   ideasRoutes,
